# api/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0
  contact:
    name: Source code
    url: https://github.com/Alexander-D-Karpov/calendar
  description: |
    JSON API for calendars, events, todos and share links. All operations live under `/api/v1`.

    ## Authentication

    Send an API token in the `Authorization` header:

        Authorization: Bearer cal_<prefix>_<secret>

    Create tokens in [Settings → API tokens](/settings/tokens). A token is shown once, right after it is created, and can be revoked at any time. To try requests on this page with a token, press **Authorize** and paste it.

    Requests from a signed-in browser session work too. With a session, unsafe methods (`POST`, `PUT`, `PATCH` and `DELETE`) also need the `X-CSRF-Token` header. **Try it out** on this page adds that header when you are signed in.

    ## Scopes

    Each token carries one or more scopes, and each operation names the scope it needs. A `write` scope includes `read` access to the same resource. Browser sessions have every scope.

    | Scope | Allows |
    |---|---|
    | `account:read` | Reading the account profile |
    | `account:write` | Changing account settings and the sleep schedule |
    | `calendars:read` | Reading calendars and events |
    | `calendars:write` | Creating, changing and deleting calendars and events |
    | `todos:read` | Reading todo lists, todos and checks |
    | `todos:write` | Creating, changing, completing and deleting todos |
...

...

    ## MCP

...

...
servers:
...
paths:
...
  /calendars:
    get:
      tags: [Calendars]
      operationId: listCalendars
      summary: List calendars
      description: |
        Returns every calendar of the account ordered by `position`. Hidden calendars are included.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      responses:
        "200":
          description: All calendars of the account.
          headers: &okHeaders
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarList"
              example:
                items:
                  - &calendarExample
                    id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    name: Work
                    color: "#3b6ea5"
                    description: ""
                    timezone: Europe/Moscow
                    kind: local
                    read_only: false
                    is_default: false
                    hidden: false
                    position: 1
                    default_reminders: [10]
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T14:02:11Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Calendars]
      operationId: createCalendar
      summary: Create a calendar
      description: |
        Creates a local calendar. Only `name` is required. The calendar is placed after the others unless `position` is set. An account can have at most 100 calendars.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              name: Work
              color: "#3b6ea5"
              timezone: Europe/Moscow
              default_reminders: [10]
      responses:
        "201":
          description: The created calendar.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /calendars/{id}:
    parameters:
      - $ref: "#/components/parameters/CalendarID"
    get:
      tags: [Calendars]
      operationId: getCalendar
      summary: Get a calendar
      description: |
        Returns one calendar. Send the `ETag` from an earlier response in `If-None-Match` to get `304` when nothing changed.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The calendar.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags: [Calendars]
      operationId: updateCalendar
      summary: Update a calendar
      description: |
        Changes the fields present in the body as a JSON Merge Patch. Send `"timezone": null` to fall back to the account timezone. `kind`, `read_only` and `is_default` cannot be changed.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              name: Personal
              timezone: null
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarInput"
            example:
              hidden: true
      responses:
        "200":
          description: The updated calendar.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
              example: *calendarExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Calendars]
      operationId: deleteCalendar
      summary: Delete a calendar
      description: |
        Deletes the calendar together with all of its events and removes it from share links. The default calendar cannot be deleted.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: The calendar and its events were deleted.
          headers: *okHeaders
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /events:
    get:
      tags: [Events]
      operationId: listEvents
      summary: List events in a time range
      description: |
        Returns events that overlap the range from `from` to `to`, ordered by start. The range can be at most 400 days. All-day events are matched by date in the account timezone.

        Recurring series are expanded into occurrences. An occurrence of a series carries the series `id` and its original start in `instance`. An occurrence changed on its own has its own `id`, the series in `series_id` and its original start in `instance`. With `expand=false` the stored series, changed occurrences and single events are returned without expansion.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/RangeFrom"
        - $ref: "#/components/parameters/RangeTo"
        - $ref: "#/components/parameters/CalendarFilter"
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: Events and occurrences in the range.
          headers: *okHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
              example:
                items:
                  - &eventExample
                    id: "0191f2c5-1b2c-7d4e-8f90-a1b2c3d4e5f6"
                    calendar_id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    uid: "0191f2c5-1b2c-7d4e-8f90-a1b2c3d4e5f6"
                    series_id: null
                    instance: "2026-09-11T10:00:00+03:00"
                    recurring: true
                    title: Standup
                    body: "Agenda and **notes**"
                    location: Office
                    url: ""
                    all_day: false
                    start: "2026-09-11T10:00:00+03:00"
                    end: "2026-09-11T10:15:00+03:00"
                    timezone: Europe/Moscow
                    rrule: "FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR"
                    rdate: []
                    exdate: ["2026-09-14T10:00:00+03:00"]
                    status: confirmed
                    transparency: opaque
                    visibility: default
                    color: null
                    read_only: false
                    reminders: [10]
                    version: 3
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T14:02:11Z"
                  - id: "0191f2c6-2d3e-7f40-8a51-b2c3d4e5f607"
                    calendar_id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    uid: "0191f2c6-2d3e-7f40-8a51-b2c3d4e5f607"
                    series_id: null
                    instance: null
                    recurring: false
                    title: Conference
                    body: ""
                    location: Vilnius
                    url: "https://example.com/conf"
                    all_day: true
                    start: "2026-09-20"
                    end: "2026-09-22"
                    timezone: null
                    rrule: null
                    rdate: []
                    exdate: []
                    status: tentative
                    transparency: transparent
                    visibility: private
                    color: "#aa3355"
                    read_only: false
                    reminders: []
                    version: 1
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T13:56:42Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Events]
      operationId: createEvent
      summary: Create an event
      description: |
        Creates an event. Only `start` is required. Without `calendar_id` the event goes to the default calendar. Without `end` a timed event lasts one hour and an all-day event one day. Times without an offset are read in `timezone`, which defaults to the calendar timezone and then the account timezone. Without `reminders` the calendar's default reminders are used.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
            examples:
              recurring:
                summary: Weekday standup
                value:
                  title: Standup
                  start: "2026-09-11T10:00:00+03:00"
                  end: "2026-09-11T10:15:00+03:00"
                  timezone: Europe/Moscow
                  rrule: "FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR"
                  reminders: [10]
              allDay:
                summary: All-day event
                value:
                  title: Conference
                  all_day: true
                  start: "2026-09-20"
                  end: "2026-09-22"
      responses:
        "201":
          description: The created event.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /events/{id}:
    parameters:
      - $ref: "#/components/parameters/EventID"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get an event
      description: |
        Returns one stored event: a single event, a recurring series or an occurrence changed on its own. Use `GET /events/{id}/instances` for the occurrences of a series.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The event.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update an event
      description: |
        Changes the fields present in the body as a JSON Merge Patch. Changing `start` without `end` keeps the duration.

        For a recurring series, `scope` and `instance` choose what changes:

        - `this` changes one occurrence and returns it as a changed occurrence.
        - `following` ends the series before the occurrence and starts a new series there with the changes applied. The response is the new series. Its `uid` is the old one with `_R` and the occurrence start in UTC appended, as in `<uid>_R20260911T100000Z`, or a fresh ID when that would pass 1000 characters. Changed occurrences, extra dates and excluded dates at or after the split move to the new series.
        - `all` applies the fields to the series itself, so `start` sets the first occurrence. Moving the start also moves excluded dates, extra dates and changed occurrences by the same amount, unless the request sets `rdate` or `exdate` itself, or changes `all_day` or `rrule`.

        A changed occurrence that no longer falls on the rule is deleted.

        Recurrence fields can only change with `following` or `all`. `uid` cannot be changed.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/EditScope"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/EventInput"
            example:
              start: "2026-09-11T11:00:00+03:00"
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
            example:
              title: Standup moved
              location: null
      responses:
        "200":
          description: The changed event, changed occurrence or new series.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Events]
      operationId: deleteEvent
      summary: Delete an event
      description: |
        Deletes the event and moves it to the trash. For a recurring series, `this` excludes one occurrence and deletes the changed copy of it if there is one, `following` ends the series before the occurrence, and `all` deletes the series with every changed occurrence.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/EditScope"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: The event or occurrences were deleted.
          headers: *okHeaders
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /todos:
    get:
      tags: [Todos]
      operationId: listTodos
      summary: List todos
      description: |
        Returns todos in creation order, open todos by default. Subtasks are included; filter them with `parent_id`. Results come in pages of at most `limit` items. Pass `next_cursor` as `cursor` to get the next page, it is null on the last page.

        Requires scope `todos:read`.
      x-scope: todos:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/TodoListFilter"
        - $ref: "#/components/parameters/ParentFilter"
        - $ref: "#/components/parameters/TodoStatus"
        - $ref: "#/components/parameters/DueFrom"
        - $ref: "#/components/parameters/DueTo"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: One page of todos.
          headers: *okHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoPage"
              example:
                items:
                  - &todoExample
                    id: "0191f2c8-4f50-7b62-8c73-d4e5f6071829"
                    list_id: "0191f2c7-3e4f-7a51-9b62-c3d4e5f60718"
                    parent_id: null
                    title: Assemble GPU node
                    body: Check PSU rails first
                    status: needs_action
                    priority: 2
                    position: V
                    due_date: "2026-09-12"
                    due_time: "18:00"
                    duration_min: 90
                    timezone: Europe/Moscow
                    rrule: null
                    show_on_calendar: true
                    completed_at: null
                    reminders: [30]
                    checks:
                      - &checkExample
                        id: "0191f2c9-5061-7c73-9d84-e5f607182930"
                        text: Riser cables
                        done: true
                        position: V
                        done_at: "2026-09-10T14:02:11Z"
                      - id: "0191f2c9-5061-7c73-9d84-e5f607182931"
                        text: Thermal paste
                        done: false
                        position: l
                        done_at: null
                    version: 2
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T14:02:11Z"
                next_cursor: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Todos]
      operationId: createTodo
      summary: Create a todo
      description: |
        Creates a todo. Only `title` is required. Without `list_id` the todo goes to the default list, or to the list of `parent_id`. `due_time` needs `due_date`, and `timezone` defaults to the account timezone. `show_on_calendar` defaults to true when a time is set. `checks` can only be set here, change them later with the checks endpoints.

        Requires scope `todos:write`.
      x-scope: todos:write
      security: *writeAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoInput"
            example:
              title: Assemble GPU node
              due_date: "2026-09-12"
              due_time: "18:00"
              duration_min: 90
              priority: 2
              checks:
                - text: Riser cables
                - text: Thermal paste
      responses:
        "201":
          description: The created todo.
          headers: *createdHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
              example: *todoExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: cal_<prefix>_<secret>
      description: API token created in Settings → API tokens.
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: >-
        Browser session cookie, set by signing in. It is named `sid` when the server runs over plain HTTP.
        Unsafe methods also need the `X-CSRF-Token` header.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        CSRF token of the browser session, required together with `sessionCookie` on unsafe methods.
        Pages expose it in the `csrf-token` meta tag.
  parameters:
    CalendarID:
      name: id
      in: path
      required: true
      description: Calendar ID.
      schema:
        type: string
        format: uuid
      example: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
...
  headers:
    ETag:
      description: Version of the resource. Send it in `If-Match` to avoid overwriting other changes.
      schema:
        type: string
      example: '"lx3kq9zb"'
    Location:
      description: URL of the created resource.
      schema:
        type: string
      example: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
    RateLimitLimit:
      description: Requests allowed per window.
      schema:
        type: integer
        minimum: 1
      example: 600
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
        minimum: 0
      example: 599
    RateLimitReset:
      description: Seconds until the full quota is available again.
      schema:
        type: integer
        minimum: 0
      example: 1
...
    RequestID:
      description: Request identifier, also returned in error bodies as `request_id`.
      schema:
        type: string
      example: tQ6DNZRbQcyrEPPA
  responses:
...
    Unauthorized:
      description: The token or session is missing, invalid, revoked or expired.
      headers:
        WWW-Authenticate:
          description: Authentication scheme to use.
          schema:
            type: string
          example: 'Bearer realm="calendar"'
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unauthorized
            status: 401
            detail: invalid api token
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
...
  schemas:
...
    CalendarList:
      type: object
      description: All calendars of the account.
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          description: Calendars ordered by position.
          items:
            $ref: "#/components/schemas/Calendar"
...